requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
sentence-transformers
pdfplumber
faiss-cpu
